+++ NEWS.md
@@ -16,6 +16,10 @@ for human readability.
 - New time integrator `PairedExplicitRK4`, implementing the fourth-order paired explicit Runge-Kutta
   method with [Convex.jl](https://github.com/jump-dev/Convex.jl), [ECOS.jl](https://github.com/jump-dev/ECOS.jl) ([#2147])
 
+#### Changed
+
+- The required Julia version is updated to v1.10.
+
 ## Changes when updating to v0.9 from v0.8.x
 
 #### Added